Add network event stream support
