+++ src/CMakeLists.txt
@@ -103,6 +103,7 @@ if(REALM_ENABLE_SYNC)
         sync/remote_mongo_collection.hpp
         sync/remote_mongo_database.hpp
         sync/push_client.hpp
+        sync/event_stream.hpp
 
         sync/impl/sync_client.hpp
         sync/impl/sync_file.hpp
@@ -125,6 +126,7 @@ if(REALM_ENABLE_SYNC)
         sync/remote_mongo_collection.cpp
         sync/remote_mongo_database.cpp
         sync/push_client.cpp
+        sync/event_stream.cpp
         sync/impl/sync_file.cpp
         sync/impl/sync_metadata.cpp
         util/bson/bson.cpp
