+++ src/sync/remote_mongo_collection.hpp
@@ -20,10 +20,12 @@
 #define REMOTE_MONGO_COLLECTION_HPP
 
 #include "sync/app_service_client.hpp"
+#include "event_stream.hpp"
 #include <realm/util/optional.hpp>
 #include <json.hpp>
 #include <string>
 #include <vector>
+#include <set>
 
 namespace realm {
 namespace app {
@@ -299,6 +301,13 @@ public:
     void find_one_and_delete(const bson::BsonDocument& filter_bson,
                              std::function<void(util::Optional<bson::BsonDocument>, util::Optional<AppError>)> completion_block);
 
+
+    /// Watches a collection
+    /// @param ids  A `Array` that contains all the ids to watch.
+    /// @param filter_bson  A `Document` representing the match criteria.
+    /// @param completion_block The result of the attempt to watch a document.
+    void watch(const bson::BsonArray& ids, const bson::BsonDocument& filter_bson, std::function<void(util::Optional<EventStream>, util::Optional<AppError>)> completion_block);
+
     /* 
      * SDKs should also support a watch method with the following 3 overloads:
      *      watch()
@@ -341,97 +350,6 @@ private:
     std::shared_ptr<AppServiceClient> m_service;
 
     std::string m_service_name;
-};
-
-/**
- * Simplifies the handling the stream for collection.watch() API.
- * 
- * General pattern for languages with pull-based async generators (preferred):
- *    auto request = app.make_streaming_request("watch", ...);
- *    auto reply = await doHttpRequestUsingNativeLibs(request);
- *    if (reply.error)
- *        throw reply.error;
- *    auto ws = WatchStream();
- *    for await (chunk : reply.body) {
- *        ws.feedBuffer(chunk);
- *        while (ws.state == WatchStream::HAVE_EVENT) {
- *            yield ws.nextEvent();
- *        }
- *        if (ws.state == WatchStream::HAVE_ERROR)
- *            throw ws.error;
- *    }
- *
- * General pattern for languages with only push-based streams:
- *    auto request = app.make_streaming_request("watch", ...);
- *    doHttpRequestUsingNativeLibs(request, {
- *        .onError = [downstream](error) { downstream.onError(error); },
- *        .onHeadersDone = [downstream](reply) {
- *            if (reply.error) 
- *                downstream.onError(error);
- *        },
- *        .onBodyChunk = [downstream, ws = WatchStream()](chunk) {
- *            ws.feedBuffer(chunk);
- *            while (ws.state == WatchStream::HAVE_EVENT) {
- *                downstream.nextEvent(ws.nextEvent());
- *            }
- *            if (ws.state == WatchStream::HAVE_ERROR)
- *                downstream.onError(ws.error);
- *        }
- *    });
- */
-struct WatchStream {
-    // NOTE: this is a fully processed event, not a single "data: foo" line!
-    struct ServerSentEvent {
-        std::string_view data;
-        std::string_view eventType = "message";
-    };
-
-    // Call these when you have data, in whatever shape is easiest for your SDK to get.
-    // Pick one, mixing and matching on a single instance isn't supported.
-    // These can only be called in NEED_DATA state, which is the initial state.
-    void feed_buffer(std::string_view); // May have multiple and/or partial lines.
-    void feed_line(std::string_view); // May include terminating CR and/or LF (not required).
-    void feed_sse(ServerSentEvent); // Only interested in "message" and "error" events. Others are ignored.
-
-    // Call state() to see what to do next.
-    enum State {
-        NEED_DATA, // Need to call one of the feed functions.
-        HAVE_EVENT, // Call next_event() to consume an event.
-        HAVE_ERROR, // Call error().
-    };
-    State state() const { return m_state; }
-
-    // Consumes the returned event. If you used feed_buffer(), there may be another event or error after this one,
-    // so you need to call state() again to see what to do next.
-    bson::BsonDocument next_event() {
-        REALM_ASSERT(m_state == HAVE_EVENT);
-        auto out = std::move(m_next_event);
-        m_state = NEED_DATA;
-        advance_buffer_state();
-        return out;
-    }
-
-    // Once this enters the error state, it stays that way. You should not feed any more data.
-    const app::AppError& error() const {
-        REALM_ASSERT(m_state == HAVE_ERROR);
-        return *m_error;
-    }
-
-private:
-    void advance_buffer_state();
-
-    State m_state = NEED_DATA;
-    util::Optional<app::AppError> m_error;
-    bson::BsonDocument m_next_event;
-
-    // Used by feed_buffer to construct lines
-    std::string m_buffer;
-    size_t m_buffer_offset = 0;
-
-    // Used by feed_line for building the next SSE
-    std::string m_event_type;
-    std::string m_data_buffer;
-
 };
 
 } // namespace app
